fix increase unittest

+++ python/paddle/tensor/stat.py
@@ -712,7 +712,7 @@ def _compute_quantile(
         if interpolation == "midpoint":
             return (tensor_upper + tensor_below) / 2
 
-        weights = (index - indices_below).astype(x.dtype)
+        weights = (index - indices_below.astype(index.dtype)).astype(x.dtype)
         # "linear"
         return paddle.lerp(
             tensor_below.astype(x.dtype),
